Initial commit
This provides the foundation for building Jailhouse demo images. For now
we support building an x86-64 image to be executed in a virtual machine
(QEMU/KVM). It uses the Isar build system for Debian, at this point from
a staging repository that is intended to be merged into upstream.

We are currently building Jailhouse release 0.8 and kernel 4.14.18 with
Jailhouse patches applied.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>