Add support for building the latest version
Call "build-images --latest", and the kas-latest.yml will be selected.
That switches the preferred version of Jailhouse to 0.9999-next which
pulls from the next branch of Jailhouse.

As the configs directory in that version was restructured, we need to
adjust jailhouse.install to cater both setups. Also prepare it for
non-x86 demos at this chance.

This can be extended later on to also pulling the latest Jailhouse
kernel branch.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>