customizations: Enable outgoing network
We are running the e1000e as NIC to reach the outer world. Configure it
to dhcp so that it gets an IP from QEMU.

Unfortunately, MSI-X is broken with that devices model and recent kernel
drivers, including 4.14. Add a modprobe rule to switch the driver back
to INTx as workaround.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>