linux-jailhouse: Add support for Orange Pi Zero
This brings the defconfig and a bunch of kernel patches to enable the
Orange Pi Zero for the Jailhouse 4.14 kernel. The patches are mostly
related to enabled the Ethernet interface of the board and are backports
for the 4.17 kernel. One patch adds the required bits to the device tree
for using an out-of-tree driver for the WLAN interface.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>