customizations: Add support for Orange Pi Zero
Add the bash_history file and extend the ivshmem-net configuration
because the Orange Pi has no PCI, thus we end up with the virtual NIC on
domain 0.

Futhermore, install the u-boot-script package from Isar so that U-Boot
will find the necessary boot information. We need to customize its
config with an additional U-Boot command so that the FDT is not
relocated but used in-place. We do not need the Debian initrd for
booting, and as we are touching the config, inject our kernel command
line this way.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>