build-images: Add image selection menu
We are starting to pile up images, and the user will rarely like to
build them all or select one via the unhandle KAS_TARGET variable. Add
an interactive selection mode to build-images.sh which presents all
available images and asks the user to select one or more.

The image list is maintained in an additional file which allows to add
some descriptions. File format is simple: <MACHINE><TABS><DESCRIPTION>.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>