Hook up Intel NUC6CAY support
This add the last missing pieces to support the Apollo Lake variant of
the Intel NUCs with up to 8GB of RAM. As noted when adding the Jailhouse
bits, this is limited to --latest at this point because of missing
relevant commits in the 0.9.1 release.

The NUC6CAY comes without a UART interface. Therefore, we are using the
new EFI framebuffer console and, of course, intensively the virtual
console. All the rest is standard again.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>