u-boot: Add recipe to build MACCHIATObin bootloader from source
This allows to switch to upstream U-Boot and ATF, resolving all
shortcomings of the pre-installed downstream version on the board. We
just need to install the generated binary on an SD card at the right
location and switch booting to that medium.

The build is complex and requires a little fix in ATF. But by deriving
from u-boot-custom.inc, we can widely focus on specialties when bringing
5(!) different sources together to form the bootable artifact.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>