Introduce machine configs
This allows to select a target also without using multiconfig, which
will be needed for this layer with the upcoming Isar version. The
multiconfigs are now only selecting the machine-distro tuple.

To avoid colliding with the upstream machine/qemuamd/arm64.conf files
which can lead to surprising results, depending on the BBPATH entry
ordering, rename those machines to qemu-amd/arm64.

This reveals the issue that we were silently relying on upstream machine
configs so far for the dependency run of bitbake. Now that we renamed
the qemu machines, this no longer works, and we have to set explicit
machine and distro values in kas.yml. Switch to multiconfig-free build
mode for the default target at this chance. This will not affect the
build-images.sh, though.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>