linux-jailhouse: Disable CONFIG_ISA_DMA_API and update config
This feature will become a conflict with using the kernel in the
non-root cell from 4.18 onward. Moreover, no platform we need to support
needs or even has ISA DMA. So already turn it off in the current config.
This requires to enable CONFIG_EXPERT while ensuring that none of the
suddenly available debug options get turned on.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>