non-root-initramfs: Rework rules
This primarily prepares to new Isar where the build is executed
unprivileged. This causes dpkg-buildpackage to call the rules under
fakeroot which conflicts with the use by buildroot itself. Avoid this by
moving the compilation into the build target which is not fakeroot-
affected.

Furthermore, avoid that dh saves and restores autotools artifacts across
clean. This is not needed, and it can break because builroot removes the
target directories during distclean.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>