Update to Debian buster
Buster brings gcc 8, and that means we need the related patch for
Jailhouse 0.11.

The non-root-initramfs has to be switches to newer rules format that
allows to disable fakeroot reliably (to avoid the nesting issue).

And the orangepi target can drop the u-boot backport pick - injecting
additional apt sources via the machine conf was broken anyway and would
have exploded when adding a second ARM target.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>