u-boot-ultra96: Import fsbl and bitstream binaries
Our public source, an OpenSUSE contrib rpm, is no longer available for
direct download. So it's no longer avoidable to adopt the two generated/
pre-built binaries we need to compile boot.bin. They originate from

https://build.opensuse.org/package/show/devel:ARM:Factory:Contrib:Zynq:ZCU100/zynqmp-hdf
https://build.opensuse.org/package/show/devel:ARM:Factory:Contrib:Zynq:ZCU100/zynqmp-fsbl

That's a bitstream for the Ultra96 which enables the console UART and a
fitting first-stage bootloader. Both are originally generated by the
proprietary Xilinx toolchain. The fsbl is then widely available in
source code, but the some binary-only bit require a specific, old
toolchain to link against. Trying that path would likely end up in
pulling most of petalinux.

Yes, this is a huge mess. We just didn't invent it.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>