non-root-initramfs: Include ivshmem-demo from Jailhouse
Pull the ivshmem-demo code from current Jailhouse, build it for the
targets and install it in the initramfs images. This allows to test
ivshmem in non-root Linux cells.

This recipe change is avoiding a direct dependency on the Jailhouse
package for obtaining the source file so that changes to the latter do
not start to trigger the lengthy non-root-initramfs build. There aren't
many changes to the demo application in Jailhouse expected, so this
should be fine.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>