linux-jailhouse: Add RPi4 variant of the kernel
The Raspberry Pi 4 is still on its way towards mainline, probably the
next LTS will look much brighter. For now, we need to maintain a
separate branch and provide a separate recipe here in order to have a
functioning kernel with the required Jailhouse patches.

The defconfig is derived from bcm2711_defconfig in that downstream
kernel.

We do not have an -rt variant at the moment because the RPi hackers
didn't provide one for 5.4 yet. So, building with --rt will just fail.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>