Hook up Raspberry Pi 4 support
Add the machine, multiconfig and wks file, extend the images.list.

The image is assembled very similarly to Raspbian, just focusing on the
support for the model 4. It is using the Raspberry bootloader to start
Linux directly but pulls in ATF before that. This implies that Linux
kernel updates on a target would need another manual step to update also
kernel.img in the boot partition.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>