jailhouse: Update to 0.12
We have a new stable version, switch to it as default. As usual, this
allows to drop merged patches, and we can also remove the deviating
NUC6CAY configs for 0.11.

Unfortunately, 0.12 had a regression in some configs that affects
supported targets. So we need to add a new patch from the beginning.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>