crosstool-ng-microblaze: Switch to github download
crosstool-ng.org is down for a couple of days. Better us the more
reliable github source. This just requires some minor adjustments to
paths and the rules file because github packages a git snapshot, not a
real release package. The snapshot also already contains a debian
folder, but we better rely on our version that is tuned for the
microblaze use case - and known to work.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>