linux-jailhouse: Add patches for Ultra96-v2
The SDIO-attached Microchip ATWILC3000 on the Ultra96-v2 requires a
number of tweaks to sdhci-of-arasan and the zynqmp firmware drivers that
only made it into later kernels or are not yet queued for upstream.
Those are the majority of additions. They have been cherry-picked and
adjusted where needed from linux-xlnx 2020.01.

The final patch adds the rev2 device tree which tunes to the specifics
of that board, compared to rev1. This cannot go upstream yet as wilc3000
is no official binding so far.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>