u-boot: Add specifics for Ultra96-v2
Those are the bitstream and the fsbl, both generated from pynq-ultra96
3e71aa7772e3.

Any attempt to rebuild the fsbl within Debian failed so far as there is
no aarch64-none-elf-gcc available, and generating that for this thing is
not worth with effort. The Xilinx embeddedsw should rather be improved
to build the fsbl with standard aarch64-linux-gnu-gcc, just like most
other firmware components do.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>