wilc: Add out-of-tree wifi driver for Ultra96-v2
The v2 has a different driver than the v1, which apparently had
regulatory reasons. The big downside is that the driver is not yet
mainline. Pick up the fork that Avnet generated for the Ultra96-v2 for
now. This should be clean up later by moving to the original downstream
version or to upstream once the wilc1000 support there is expanded.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>