macchiatobin: Convert to stand-alone TF-A recipe and update versions
Rework the U-Boot recipes to use TF-A as a package, rather than building
it in an embedded fashion.

This means turning things around: U-Boot is built first as dependency of
TF-A. IMAGER_INSTALL and IMAGER_BUILD_DEPS are updated to pull TF-A.
And the TF-A artifacts mv-ddr-marvell and binaries-marvell are moved to
the new stand-alone recipe.

While at it, bump to TF-A 2.3 and U-Boot 2020.10. TF-A needs one patch
backported from the development tree in order to fix parallel build for
the a80x0_mcbin platform.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>