Update to kas 2.4
kas-docker is now kas-container. Use the chance of the update to
embedded kas-container, rather than downloading it on demand which comes
with the risk of running from an older version.

With the update, building on an arm64 host is now possible as well,
though this wasn't tried for jailhouse-images yet.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>