linux-jailhouse[-rt], wilc: Switch to 5.10
This allows to drop a lot of ZynqMP-related patches. Still, two new ones
are needed due to a regression of the DTs for Ultra96 v1 and v2, see
also https://lkml.org/lkml/2021/3/27/354.

The downstream wilc driver for the Ultra96v2 requires a backport from
upstream in order to work with the newer kernel. For simplicity reasons,
that patch will not account for older kernels anymore, thus is part of
this commit.

RPi4 will be handled separately.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>