nuc6cay, ipc127e, hikey620: Fix wks files for newer Isar versions
Since the wic update, the boot partition is actually tried to be
mounted. This fails due CONFIG_SYSFS_DEPRECATED - systemd simply does
not find the device node.

We also need to move the mount point to avoid overriding the content in
/boot. Finally, --use-uuid permits to use it irrespective of the boot
medium.

Also drop the unused --ondisk parameters at this chance.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>