customizations: include wildcard into SRC_URI
That kind of tells bitbake that the file is "optional" and it will not
warn about it missing when parsing the recipes. It would find it missing
in the install task.

That allows re-using jailhouse-images in projects that do not even
install the customizations package and would receive warnings when not
having a config for the package and their machine.

Signed-off-by: Henning Schild <henning.schild@siemens.com>
[Jan: match more exactly]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>