linux-jailhouse: Update to 5.15
Move to the next LTS kernel. This allows to drop 2 fix-ups for Ultra96.

Along the kernel update, the RPi4 alias overlay needs to be adjusted as
the kernel renamed nodes in the target device tree.

Furthermore, we need to disable CONFIG_SUSPEND on x86 for the sake of
the non-root Linux cell. Otherwise, an unsupported access to
MSR_AMD64_DE_CFG would be triggered on Intel CPUs.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>